Convert systematic handling to log-normal
I've edited the lines that handle the systematics (and particularly the combination of the systematics to get the total) to make them behave appropriately for log-normal nuisance parameters.

Please check the "symmetric error" portion circa line 475.  I am not sure whether `error` is the shifted yield, the difference between the shifted yield and the central yield, the ratio of the shifted yield to the central yield, or the ratio of the difference between the shifted yield and the central yield to the central yield.  The code I wrote assumes that it is the ratio of the shifted yield to the central yield, just like `syst_up` and `syst_down` in lines 452 and 453.  If it is something else, then you'll need to convert it to that ratio.